FPT_Backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
alembic
fastapi-mail 
bcrypt<4.0.0
passlib